moved the components into the components folder
